+++ examples/Program.cs
@@ -13,6 +13,9 @@ internal class Program
         // Get OpenAI API key from environment variable
         var apiKey = Environment.GetEnvironmentVariable("LLM_API_KEY") ?? throw new InvalidOperationException("Set LLM_API_KEY env var.");
 
+        Console.WriteLine("---------------------------- SIMPLE STREAMING TEST --------------------------");
+        await SimpleStreamingTest.RunAsync(apiKey);
+
         Console.WriteLine("---------------------------- RE/ACT EXAMPLE --------------------------");
         await ReactExample.RunAsync(apiKey);
 
@@ -22,7 +25,5 @@ internal class Program
         Console.WriteLine("---------------------------- TREE OF THOUGHTS EXAMPLE --------------------------");
         await TreeOfThoughsExample.RunAsync(apiKey);
 
-        Console.WriteLine("---------------------------- SIMPLE STREAMING TEST --------------------------");
-        await SimpleStreamingTest.RunAsync(apiKey);
     }
 }